requirements.txt
annotated-types==0.7.0
anyio==4.12.0
boto3==1.42.5
botocore==1.42.5
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
fastapi==0.100.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hypercorn==0.14.4
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
priority==2.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
requests==2.32.5
resend==2.19.0
s3transfer==0.16.0
six==1.17.0
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
wsproto==1.3.2
